Completed LicensePlate Homework and also did CharDemo in class
I completed the LicensePlate homework. In this program, users enter a license plate and then it will evaluate the input as VALID or INVALID depending on if it passes the requirements. They are listed in the program.

CharDemo went over strings and ways you can check for the characters within the strings.