Added and completed RandomMusicPlaylist-V3
This is another redone version of the RandomMusicPlaylist where it now gives users options. It will then either quit or save the generated playlists and let the user name it, finally storing it in a new folder called "playlists".

This was a good challenge, trying to figure out the directories was the hardest part for me!