Completed MathGame
Completed MathGame Homework.

This program will provide a game for the user. The user can select the difficulty [1-3] and the number of questions [3-10]. It will then prompt the user with the questions, they have three attempts to get it right, if they don't, it moves on the the next question. After completing the questions it will then calculate the score based on 0%-100%. Then the program ends. There is properly error handling and try catch blocks for exceptions. This was a very fun challenge that I really enjoyed working on!