Started and Completed AnalyzeMusicPlaylist Homework
The program is to receive the path to a tab-delimited data file containing music playlist information as a command line argument. If the tab-delimited data file is in the same directory as the program, then the user can just supply the name of the file as the path.

The records in the tab-delimited data file are to be read by the program and placed in a collection of structured data. Then, Language INtegrated Query (LINQ) along with general C# code is to be used to determine the answers to a set of questions and a report is to be generated in a text file that is saved. The path of the report file is also to be supplied as a command line argument. If the path of the report file is just the name of the file, then it will be saved in the current directory.