Started FinalProject.cs
In this final project I will create an online banking application for a new bank. A new bank is opening in town and I have been tasked with creating a mock-up of an online banking application. The application will load customer data from a .csv file and allow customers to log in and access a number of services like depositing, withdrawals, transfers, etc. The customer data file should be updated as account services are used. There will also be an administrator login section for bank employees to log in and view different reports. I was also given an employee data file with employee information.

This is about four days worth of coding put together!